IDG-Apollo-4611 [Task] add the vin in ch vehicle

+++ modules/canbus/vehicle/ch/protocol/brake_command_111.cc
@@ -29,7 +29,7 @@ const int32_t Brakecommand111::ID = 0x111;
 Brakecommand111::Brakecommand111() { Reset(); }
 
 uint32_t Brakecommand111::GetPeriod() const {
-  // modify every protocol's period manually
+  // TODO(All) :  modify every protocol's period manually
   static const uint32_t PERIOD = 20 * 1000;
   return PERIOD;
 }
@@ -40,7 +40,7 @@ void Brakecommand111::UpdateData(uint8_t* data) {
 }
 
 void Brakecommand111::Reset() {
-  // you should check this manually
+  // TODO(All) :  you should check this manually
   brake_pedal_en_ctrl_ = Brake_command_111::BRAKE_PEDAL_EN_CTRL_DISABLE;
   brake_pedal_cmd_ = 0;
 }
@@ -51,18 +51,18 @@ Brakecommand111* Brakecommand111::set_brake_pedal_en_ctrl(
   return this;
 }
 
-// config detail: {'description': 'brake pedal enable bit(Command)', 'enum': {0:
-// 'BRAKE_PEDAL_EN_CTRL_DISABLE', 1: 'BRAKE_PEDAL_EN_CTRL_ENABLE'},
-// 'precision': 1.0, 'len': 8, 'name': 'BRAKE_PEDAL_EN_CTRL', 'is_signed_var':
-// False, 'offset': 0.0, 'physical_range': '[0|1]', 'bit': 0, 'type': 'enum',
-// 'order': 'intel', 'physical_unit': ''}
+// config detail: {'bit': 0, 'description': 'brake pedal enable bit(Command)',
+// 'enum': {0: 'BRAKE_PEDAL_EN_CTRL_DISABLE', 1: 'BRAKE_PEDAL_EN_CTRL_ENABLE'},
+// 'is_signed_var': False, 'len': 8, 'name': 'BRAKE_PEDAL_EN_CTRL', 'offset':
+// 0.0, 'order': 'intel', 'physical_range': '[0|1]', 'physical_unit': '',
+// 'precision': 1.0, 'type': 'enum'}
 void Brakecommand111::set_p_brake_pedal_en_ctrl(
     uint8_t* data,
     Brake_command_111::Brake_pedal_en_ctrlType brake_pedal_en_ctrl) {
   int x = brake_pedal_en_ctrl;
 
   Byte to_set(data + 0);
-  to_set.set_value(static_cast<uint8_t>(x), 0, 8);
+  to_set.set_value(x, 0, 8);
 }
 
 Brakecommand111* Brakecommand111::set_brake_pedal_cmd(int brake_pedal_cmd) {
@@ -70,17 +70,17 @@ Brakecommand111* Brakecommand111::set_brake_pedal_cmd(int brake_pedal_cmd) {
   return this;
 }
 
-// config detail: {'description': 'Percentage of brake pedal(Command)',
-// 'offset': 0.0, 'precision': 1.0, 'len': 8, 'name': 'BRAKE_PEDAL_CMD',
-// 'is_signed_var': False, 'physical_range': '[0|100]', 'bit': 8, 'type': 'int',
-// 'order': 'intel', 'physical_unit': '%'}
+// config detail: {'bit': 8, 'description': 'Percentage of brake
+// pedal(Command)', 'is_signed_var': False, 'len': 8, 'name': 'BRAKE_PEDAL_CMD',
+// 'offset': 0.0, 'order': 'intel', 'physical_range': '[0|100]',
+// 'physical_unit': '%', 'precision': 1.0, 'type': 'int'}
 void Brakecommand111::set_p_brake_pedal_cmd(uint8_t* data,
                                             int brake_pedal_cmd) {
   brake_pedal_cmd = ProtocolData::BoundedValue(0, 100, brake_pedal_cmd);
   int x = brake_pedal_cmd;
 
   Byte to_set(data + 1);
-  to_set.set_value(static_cast<uint8_t>(x), 0, 8);
+  to_set.set_value(x, 0, 8);
 }
 
 }  // namespace ch
