+++ modules/canbus/vehicle/ch/protocol/throttle_command_110.cc
@@ -29,7 +29,7 @@ const int32_t Throttlecommand110::ID = 0x110;
 Throttlecommand110::Throttlecommand110() { Reset(); }
 
 uint32_t Throttlecommand110::GetPeriod() const {
-  // modify every protocol's period manually
+  // TODO(All) :  modify every protocol's period manually
   static const uint32_t PERIOD = 20 * 1000;
   return PERIOD;
 }
@@ -40,7 +40,7 @@ void Throttlecommand110::UpdateData(uint8_t* data) {
 }
 
 void Throttlecommand110::Reset() {
-  // you should check this manually
+  // TODO(All) :  you should check this manually
   throttle_pedal_en_ctrl_ =
       Throttle_command_110::THROTTLE_PEDAL_EN_CTRL_DISABLE;
   throttle_pedal_cmd_ = 0;
@@ -52,18 +52,18 @@ Throttlecommand110* Throttlecommand110::set_throttle_pedal_en_ctrl(
   return this;
 }
 
-// config detail: {'description': 'throttle pedal enable bit(Command)', 'enum':
-// {0: 'THROTTLE_PEDAL_EN_CTRL_DISABLE', 1: 'THROTTLE_PEDAL_EN_CTRL_ENABLE'},
-// 'precision': 1.0, 'len': 8, 'name': 'THROTTLE_PEDAL_EN_CTRL',
-// 'is_signed_var': False, 'offset': 0.0, 'physical_range': '[0|1]', 'bit': 0,
-// 'type': 'enum', 'order': 'intel', 'physical_unit': ''}
+// config detail: {'bit': 0, 'description': 'throttle pedal enable
+// bit(Command)', 'enum': {0: 'THROTTLE_PEDAL_EN_CTRL_DISABLE', 1:
+// 'THROTTLE_PEDAL_EN_CTRL_ENABLE'}, 'is_signed_var': False, 'len': 8, 'name':
+// 'THROTTLE_PEDAL_EN_CTRL', 'offset': 0.0, 'order': 'intel', 'physical_range':
+// '[0|1]', 'physical_unit': '', 'precision': 1.0, 'type': 'enum'}
 void Throttlecommand110::set_p_throttle_pedal_en_ctrl(
     uint8_t* data,
     Throttle_command_110::Throttle_pedal_en_ctrlType throttle_pedal_en_ctrl) {
   int x = throttle_pedal_en_ctrl;
 
   Byte to_set(data + 0);
-  to_set.set_value(static_cast<uint8_t>(x), 0, 8);
+  to_set.set_value(x, 0, 8);
 }
 
 Throttlecommand110* Throttlecommand110::set_throttle_pedal_cmd(
@@ -72,17 +72,17 @@ Throttlecommand110* Throttlecommand110::set_throttle_pedal_cmd(
   return this;
 }
 
-// config detail: {'description': 'Percentage of throttle pedal(Command)',
-// 'offset': 0.0, 'precision': 1.0, 'len': 8, 'name': 'THROTTLE_PEDAL_CMD',
-// 'is_signed_var': False, 'physical_range': '[0|100]', 'bit': 8, 'type': 'int',
-// 'order': 'intel', 'physical_unit': '%'}
+// config detail: {'bit': 8, 'description': 'Percentage of throttle
+// pedal(Command)', 'is_signed_var': False, 'len': 8, 'name':
+// 'THROTTLE_PEDAL_CMD', 'offset': 0.0, 'order': 'intel', 'physical_range':
+// '[0|100]', 'physical_unit': '%', 'precision': 1.0, 'type': 'int'}
 void Throttlecommand110::set_p_throttle_pedal_cmd(uint8_t* data,
                                                   int throttle_pedal_cmd) {
   throttle_pedal_cmd = ProtocolData::BoundedValue(0, 100, throttle_pedal_cmd);
   int x = throttle_pedal_cmd;
 
   Byte to_set(data + 1);
-  to_set.set_value(static_cast<uint8_t>(x), 0, 8);
+  to_set.set_value(x, 0, 8);
 }
 
 }  // namespace ch
